requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
click==8.1.4
colorama==0.4.6
distlib==0.3.6
dj-database-url==2.0.0
Django==4.1
djangorestframework==3.14.0
exceptiongroup==1.1.2
fastapi==0.100.0
FastAPI-SQLAlchemy==0.2.1
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Mako==1.2.4
MarkupSafe==2.1.3
Pillow==10.0.0
pipenv==2023.7.4
platformdirs==3.8.1
psycopg2==2.8.6
pydantic==2.0.2
pydantic_core==2.1.2
python-dotenv==1.0.0
pytz==2023.3
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.22.0
virtualenv==20.23.1
virtualenv-clone==0.5.7
zipp==3.15.0